improvement(api): send action type to cloud
There's a bit of history here.

Garden has a concept of action kinds and action types, both of which we want to store in display in
the dashboard UI.

Due to some legacy issues, were confusing the two and the API didn't support payload with the
action type set correctly.

This has since been fixed in the API but we never updated the Core payload until now.

However, there are older instances of the API still running that haven't been updated to support
this payload yet, so we feature flag it for now. We should be able to move the feature flag very
soon.